+++ lib/features/voucher/presentation/voucher_providers.dart
@@ -2,6 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_riverpod/legacy.dart';
 
 import '../../../core/di/providers.dart';
+import '../domain/entities/ticket_template.dart';
 import '../domain/entities/voucher_encoding_settings.dart';
 import '../domain/entities/voucher_entity.dart';
 
@@ -17,4 +18,9 @@ final voucherHistoryProvider = FutureProvider.autoDispose
 final voucherEncodingProvider =
     FutureProvider.autoDispose<VoucherEncodingSettings>((ref) {
       return ref.watch(voucherEncodingSettingsServiceProvider).load();
+    });
+
+final voucherTicketTemplateProvider =
+    FutureProvider.autoDispose<TicketTemplate>((ref) {
+      return ref.watch(ticketTemplateSettingsServiceProvider).loadSelected();
     });